Fix: really use `double` in `examples/drank*`.
Original commit message in repo xenia:rheinfall (rev 144)
   revno: 144
   committer: Riccardo Murri <riccardo.murri@gmail.com>
   branch nick: rheinfall
   timestamp: Sun 2010-11-14 18:10:23 +0100

+++ trunk/src.c++/examples/drank.cpp
@@ -29,7 +29,7 @@
 
 #include <config.h>
 
-typedef long long val_t;
+typedef double val_t;
 typedef long coord_t;
 
 // use WITH_MPI_SERIALIZED / WITH_MPI_MULTITHREADED to select which
